First working test using CollectionsTestSupport
Testing sequence initializer
