+++ Sources/BitArrayModule/BitArray/BitArrayOperations.swift
@@ -36,6 +36,16 @@ extension BitArray {
     }
   }
   
+  public mutating func remove(at: Int) {
+    self[at] = false
+    // needs work
+  }
+  
+  public mutating func removeLast() {
+    self[endIndex-1] = false
+    excess -= 1
+  }
+  
   internal func firstTrueIndex() -> Int {
     var counter = -1
     for item in storage {
