+++ Tests/BitArrayTests/BitArrayTests.swift
@@ -13,8 +13,16 @@ final class BitArrayTest: CollectionTestCase {
   
   let sizes: [Int] = _getSizes(BitArray.UNIT.bitWidth)
   
+  func testSequenceInitializer() {
+    withSomeUsefulBoolArrays("boolArray", ofSizes: sizes, ofUnitBitWidth: BitArray.UNIT.bitWidth) { boolArray in
+        let testBitArray: BitArray = BitArray(boolArray)
+        expectEqual(Array(testBitArray), boolArray)
+    }
+  }
+  
   func testAppend() {
-    withSomeUsefulBitArrays("bitArray", ofSizes: sizes, ofUnitBitWidth: BitArray.UNIT.bitWidth) { layout in
+    withSomeUsefulBoolArrays("bitArray", ofSizes: sizes, ofUnitBitWidth: BitArray.UNIT.bitWidth) { layout in
+      print(layout)
     }
   }
 }
